+++ src/Providers/ConsoleServiceProvider.php
@@ -37,6 +37,7 @@ use Nwidart\Modules\Commands\RuleMakeCommand;
 use Nwidart\Modules\Commands\SeedCommand;
 use Nwidart\Modules\Commands\SeedMakeCommand;
 use Nwidart\Modules\Commands\SetupCommand;
+use Nwidart\Modules\Commands\TestMakeCommand;
 use Nwidart\Modules\Commands\UnUseCommand;
 use Nwidart\Modules\Commands\UpdateCommand;
 use Nwidart\Modules\Commands\UseCommand;
@@ -88,6 +89,7 @@ class ConsoleServiceProvider extends ServiceProvider
         UpdateCommand::class,
         UseCommand::class,
         ResourceMakeCommand::class,
+        TestMakeCommand::class,
     ];
 
     /**